# exchange-rate-api/docs/swagger.yaml
basePath: /
definitions:
  models.ExchangeRate:
    properties:
      date:
        example: "2022-05-01T00:00:00.00Z"
        type: string
      destination:
        example: CHF
        type: string
      rate:
        example: 1.0456
        type: number
      source:
        example: USD
        type: string
    required:
    - date
    - destination
    - rate
    - source
    type: object
info:
  contact: {}
  description: Provides basic functionality for checking currency exchange rate.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Rate Exchange API
  version: "1.0"
paths:
  /check:
    get:
      description: basic healthcheck
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: healthcheck
      tags:
      - healthcheck
  /currencies:
    get:
      consumes:
      - application/json
      description: Returns list of all currencies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      summary: GetAllCurrencies
      tags:
      - currencies
  /exchange-rate:
    post:
      consumes:
      - application/json
      description: Inserts new exchange rate
      parameters:
      - description: New exchange rate to insert. Date has to be in RFC3339 format
          due to gin limitation. Time part will be ignored
        in: body
        name: newExchangeRate
        required: true
        schema:
          $ref: '#/definitions/models.ExchangeRate'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/models.ExchangeRate'
      summary: InsertExchangeRate
      tags:
      - exchange-rate
  /exchange-rate/all-from-date/{date}:
    get:
      consumes:
      - application/json
      description: Returns all exchange rates for the given date
      parameters:
      - description: Date for which exchange rates should be retrived. Date must be
          formated in YYYY-MM-DD
        in: path
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "203":
          description: Non-Authoritative Information
          schema:
            items:
              $ref: '#/definitions/models.ExchangeRate'
            type: array
      summary: GetAllExchangeRatesFromDate
      tags:
      - exchange-rate
  /exchange-rate/last:
    get:
      consumes:
      - application/json
      description: Returns most recent exchange rate  which is not null in database
        for source - destinaion currencies
      parameters:
      - description: source currency, default is USD
        in: query
        name: source
        type: string
      - description: destination, currency
        in: query
        name: destination
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ExchangeRate'
        "404":
          description: ""
      summary: GetLastExchangeRate
      tags:
      - exchange-rate
  /exchange-rate/range:
    get:
      consumes:
      - application/json
      description: Returns exchange rates for currencies in the time period
      parameters:
      - description: source currency, default is USD
        in: query
        name: source
        type: string
      - description: destination currency
        in: query
        name: destination
        required: true
        type: string
      - description: From date, inclusive, must be formated in YYYY-MM-DD
        in: query
        name: from
        required: true
        type: string
      - description: Till date, exclusive, must be formated in YYYY-MM-DD
        in: query
        name: till
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ExchangeRate'
            type: array
      summary: GetRangeExchangeRate
      tags:
      - exchange-rate
schemes:
- http
swagger: "2.0"